requirements.txt
langchain==0.3.26
fastapi==0.115.0
uvicorn==0.30.6
sentence-transformers==4.1.0
pypdf==5.6.1 
python-dotenv==1.1.0
langchain-pinecone==0.2.8 
langchain-ollama==0.1.2
langchain-community==0.3.26
pinecone==5.1.0
-e .